+++ src/index.js
@@ -6,18 +6,27 @@ import {
 } from "react-router-dom";
 import App from "./App";
 import Expenses from "./routes/expenses";
-import About from "./routes/about";
-import Invoices from "./routes/invoice";
+import Invoices from "./routes/invoices";
+import Invoice from "./routes/invioceid";
 
 const rootElement = document.getElementById("root");
 render(
   <BrowserRouter>
-    <Routes>
-      <Route path="/" element={<App />}>
-        <Route path="expenses" element={<Expenses />} />
-        <Route path="abouttt" element={<About />} />
-        <Route path="invoices" element={<Invoices />} />
-        <Route
+<Routes>
+  <Route path="/" element={<App />}>
+    <Route path="expenses" element={<Expenses />} />
+    <Route path="invoices" element={<Invoices />}>
+      <Route
+        index
+        element={
+          <main style={{ padding: "1rem" }}>
+            <p>Select an invoice</p>
+          </main>
+        }
+      />
+      <Route path=":invoiceId" element={<Invoice />} />
+    </Route>
+    <Route
       path="*"
       element={
         <main style={{ padding: "1rem" }}>
@@ -25,8 +34,8 @@ render(
         </main>
       }
     />
-      </Route>
-    </Routes>
+  </Route>
+</Routes>
   </BrowserRouter>,
   rootElement
 );
