+++ src/routes/data.js
@@ -31,6 +31,14 @@ let invoices = [
     },
   ];
   
-  export function getInvoices() {
-    return invoices;
-  }
+  // ...
+
+export function getInvoices() {
+  return invoices;
+}
+
+export function getInvoice(number) {
+  return invoices.find(
+    (invoice) => invoice.number === number
+  );
+}
